update statistic quick sort

+++ src/StatisticQuickSort.java
@@ -34,7 +34,7 @@ public class StatisticQuickSort {
 
 		for (int i = 1; i <= np; i++) {
 			for (int j = 1; j <= n; j++) {
-				double m = Math.pow(10, j);
+				int m = (int) Math.pow(10, j);
 				String temp = i + "\t" + m;
 				for (String clazz : classes) {
 					String command = "mpirun -np " + i + " java " + clazz + " " + m;
